Improve saving of dependent objects
When saving an object, the dependent objects will be saved automatically as well. If somewhere in this 'save cascade' an exception occurs, everything will be rolled back. However, the primary key and backend attributes of successfully saved objects were changed in the process, leading to BlitzDB thinking that they had been saved already in the database. Now, by introducing a save_cache variable that keeps track of all saved dependent objects we are able to reset these objects to their original state in case something goes wrong during a save.